requirements.txt
absl-py==2.0.0
accelerate==0.27.2
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.2.0
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
auto_gptq==0.7.0
Babel==2.13.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.43.0
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.4
coloredlogs==15.0.1
comm==0.1.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cuda-python==12.3.0
cudf-cu12==24.2.1
cupshelpers==1.0
cupy-cuda12x==13.0.0
cycler==0.12.1
dask==2024.1.1
dask-cudf-cu12==24.2.1
datasets==2.14.6
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
deep-sort-realtime==1.3.2
defer==1.0.6
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
distributed==2024.1.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
docker==7.0.0
docker-pycreds==0.4.0
duplicity==0.8.21
exceptiongroup==1.1.3
executing==2.0.0
fasteners==0.14.1
fastjsonschema==2.18.1
fastrlock==0.8.2
filelock==3.12.4
filterpy==1.4.5
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.2
gekko==1.0.6
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.1.0
gpt4all==1.0.12
grpcio==1.60.0
h11==0.14.0
httplib2==0.20.2
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.3
imageio==2.33.1
importlib-metadata==7.0.1
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.42.0
locket==1.0.0
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
notebook==7.0.4
notebook_shim==0.2.3
numba==0.59.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
nvtx==0.2.10
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.1.78
optimum==1.17.1
overrides==7.4.0
packaging==23.2
pandarallel==1.6.5
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
partd==1.4.1
pbr==5.8.0
peft==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydeck==0.8.1b0
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvjitlink-cu12==0.1.12
pyparsing==2.4.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
pyRFC3339==1.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
pytz==2022.1
PyWavelets==1.5.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
rapids-dask-dependency==24.2.0
referencing==0.30.2
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rmm-cu12==24.2.0
rouge==1.0.1
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.2
scikit-image==0.19.3
scikit-learn==1.4.1.post1
scipy==1.11.4
screen-resolution-extra==0.0.0
seaborn==0.13.0
SecretStorage==3.3.1
Send2Trash==1.8.2
sentencepiece==0.2.0
sentry-sdk==1.38.0
setproctitle==1.3.3
shiboken2==5.15.2.1
shiboken6==6.6.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
ssh-import-id==5.11
stack-data==0.6.3
stevedore==5.1.0
streamlit==1.28.2
streamlit-scrollable-textbox==0.0.3
supervisor==4.2.1
sympy==1.12
systemd-python==234
tblib==3.0.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.17.1
testresources==2.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tifffile==2023.12.9
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.38.1
trash-cli==0.17.1.14
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
unittest2==1.1.0
uri-template==1.3.0
urllib3==2.1.0
usb-creator==0.3.7
uvicorn==0.24.0.post1
validators==0.22.0
virtualenv==20.24.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.6
wandb==0.16.4
watchdog==3.0.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.1
whatever==0.7
widgetsnbextension==4.0.9
xdg==5
xkit==0.0.0
xxhash==3.4.1
yarl==1.9.2
zict==3.0.0
zipp==1.0.0
